private static function generateCustomerCode(): string     {         $faker = Faker::create();         $value = '';         $control = true;         while($control){             $value = $faker->unique()->regexify('[A-Z]{3}[0-9]{3}');             $control = Account::where('CustomerCode','ASDF23')->firstOr(function() {return false;});         }         return $value;
    }

Things like that happen when there is an idea but no exp.
using $faker to gerate the customer Code
Check whether the customerCode is already used by using the firstOr/ findOr if the value to check is used as ID.
while loop until the function inside firstOr return false.
return the latest generated Value.